[DATA]  refactor: Cloud folder structure/setting
