+++ .gitignore
@@ -206,3 +206,6 @@ cython_debug/
 # virtual machine crash logs, see http://www.java.com/en/download/help/error_hotspot.xml
 hs_err_pid*
 replay_pid*
+
+#credentials in gcp
+data/cloud_platform/*.json
