--- data/cloud_platform/gp_login.py
@@ -1,25 +1,0 @@
-
-
-#pip install
-
-from pydrive2.auth import GoogleAuth
-from pydrive2.drive import GoogleDrive
-
-directorio_credenciales = 'credentials_module.json'
-
-# INICIAR SESION
-def login():
-    GoogleAuth.DEFAULT_SETTINGS['client_config_file'] = directorio_credenciales
-    gauth = GoogleAuth()
-    gauth.LoadCredentialsFile(directorio_credenciales)
-    
-    if gauth.credentials is None:
-        gauth.LocalWebserverAuth(port_numbers=[8092])
-    elif gauth.access_token_expired:
-        gauth.Refresh()
-    else:
-        gauth.Authorize()
-        
-    gauth.SaveCredentialsFile(directorio_credenciales)
-    credenciales = GoogleDrive(gauth)
-    return credenciales
