+++ data/cloud_platform/settings.yaml
@@ -1,0 +1,13 @@
+client_config_backend: settings
+client_config:
+  client_id: 1090643125537-kc2g5kmugpgcmk8kcf3nroo2vv2ev5hi.apps.googleusercontent.com
+  client_secret: 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
+
+save_credentials: True
+save_credentials_backend: file
+save_credentials_file: credentials_module.json
+
+get_refresh_token: True
+
+oauth_scope:
+ - https://www.googleapis.com/auth/drive
